Moves moab on storage classes into single module.
