+++ app/services/moab_on_storage/moab_storage_service.rb
